🧹 Refactor setupHintSpinner to View extension function
Moved setupHintSpinner from AddResourceActivity to ViewExtensions as a Spinner extension function to remove unused private method warning, improve reusability, and resolve the code health issue.
